refactor:HeaderとFooterを別ファイルにした
ついでにcssも
