+++ src/js/components/block/header.js
@@ -1,0 +1,13 @@
+import "style/block/Header.css"
+function Header() {
+    return (
+      <div className="header">
+        <div>
+          <h1 className="h_title">GAcalc</h1>
+          <p className="h_sub_title">Genshin Ascensions calculator</p>
+        </div>
+      </div>
+    );
+  }
+
+  export default Header;
