+++ src/style/App.css
@@ -5,6 +5,7 @@
   #303841
   #be3144
 */
+/*レスポンシブ*/
 @media screen and (max-width: 600px) {
   .App{
     width: 100%;
@@ -21,26 +22,12 @@
     width: 60%;
   }
 }
+
+/*App*/
+
 .App{
   margin: 0 auto;
   background-color: white;
-}
-
-/*ヘッダー*/
-.header {
-  height: 10%;
-  margin: 0 auto;
-  display: flex;
-  background-color: #303841;
-}
-.header div{
-  width: 100%;
-  padding: 0.5rem;
-}
-.header h1, .header p {
-  color: white;
-  margin-top: 0.5rem;
-  margin-bottom: 0.5rem;
 }
 
 /*コンテンツ*/
@@ -75,48 +62,4 @@
 .main form{
   display: flex;
   justify-content: space-between;
-}
-
-/*フッター*/
-.footer {
-  margin: 0 auto;
-  margin-top: 2rem;
-  padding-top: 1rem;
-  padding-bottom: 1rem;
-  background-color: #303841;
-  height: 10%;
-}
-
-.footer > div{
-  margin-left: 0.5rem;
-}
-
-.footer p {
-  margin: 0;
-  margin-bottom: 1%;
-
-  color: white;
-}
-.footer .item {
-  width: 100%;
-  display: flex;
-  flex-wrap: nowrap;
-}
-
-.footer Button p{
-  color: #303841;
-}
-
-.footer .item a {
-  text-decoration: none;
-}
-.footer .item a div {
-  display: flex;
-  flex-wrap: nowrap;
-  vertical-align: middle;
-}
-.footer .item .twitter {
-  margin: auto 0;
-  height: 32px;
-  color: #00acee;
 }
