+++ src/style/block/Header.css
@@ -1,0 +1,15 @@
+.header {
+    height: 10%;
+    margin: 0 auto;
+    display: flex;
+    background-color: #303841;
+}
+.header div{
+width: 100%;
+padding: 0.5rem;
+}
+.h_title , .h_sub_title{
+color: white;
+margin-top: 0.5rem;
+margin-bottom: 0.5rem;
+}
